+++ lawa/operators/pdeoperators1d/weightedidentityoperator1d.tcc
@@ -12,8 +12,8 @@ WeightedIdentityOperator1D<T, Basis, Quad>::WeightedIdentityOperator1D(const Bas
 
 template <typename T, typename Basis, QuadratureType Quad>
 T
-WeightedIdentityOperator1D<T, Basis, Quad>::operator()(XType xtype1, int j1, int k1,
-                                                       XType xtype2, int j2, int k2) const
+WeightedIdentityOperator1D<T, Basis, Quad>::operator()(XType xtype1, int j1, long k1,
+                                                       XType xtype2, int j2, long k2) const
 {   
     return integral(j1, k1, xtype1, 0, j2, k2, xtype2, 0);
 }
