+++ lawa/preconditioners/preconditioners1d/diagonallevelpreconditioner1d.h
@@ -32,7 +32,7 @@ struct DiagonalLevelPreconditioner1D
     DiagonalLevelPreconditioner1D(void) { }
 
     T
-    operator()(XType xtype1, int j1, int k1) const;
+    operator()(XType xtype1, int j1, long k1) const;
 
     T
     operator()(const Index1D &index) const;
