+++ lawa/preconditioners/preconditioners2d/diagonalmatrixpreconditioner2d.h
@@ -34,8 +34,8 @@ class DiagonalMatrixPreconditioner2D
         DiagonalMatrixPreconditioner2D(const BilinearForm &a);
 
         T
-        operator()(XType xtype1, int j1, int k1,
-                   XType xtype2, int j2, int k2) const;
+        operator()(XType xtype1, int j1, long k1,
+                   XType xtype2, int j2, long k2) const;
 
         T
         operator()(const Index2D &index) const;
