+++ lawa/righthandsides/homogeneousrhs.tcc
@@ -2,7 +2,7 @@ namespace lawa {
     
 template <typename T>
 T 
-HomogeneousRHS<T>::operator()(T /*time*/, XType /*xtype*/, int /*j*/, int /*k*/) const{
+HomogeneousRHS<T>::operator()(T /*time*/, XType /*xtype*/, int /*j*/, long /*k*/) const{
     return 0;
 }
